Refactor WaterfallApplication to share code with iOS
This patch moves the interesting code out of the main waterfall.cc file
so that it can be shared with the iOS test harness.
