+++ waterfall/scenes/material_stage.h
@@ -1,0 +1,7 @@
+// Copyright 2016 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#include "escher/scene/stage.h"
+
+void InitStageForMaterial(escher::Stage* stage);
